test: split monolithic E2E test file extension.spec.ts into modular files (#260)
- Split extension.spec.ts into prompt-injection.spec.ts, drag-and-drop.spec.ts, workbench.spec.ts, card-management.spec.ts, categories.spec.ts, history-scroll.spec.ts, and settings.spec.ts.
- Verified all 17 E2E tests pass successfully in modular spec files.
- Deleted the original monolithic extension.spec.ts file.
