--- tests/e2e/extension.spec.ts
@@ -1,1463 +1,0 @@
-import { test, expect } from "@playwright/test";
-import path from "path";
-
-test.describe("Style Atelier Sandbox E2E Tests", () => {
-  test.beforeEach(async ({ page }) => {
-    page.on('console', msg => {
-      console.log(`[BROWSER CONSOLE] ${msg.type()}: ${msg.text()}`);
-    });
-    page.on('pageerror', err => {
-      console.error(`[BROWSER ERROR] ${err.message}\n${err.stack}`);
-    });
-    page.on('requestfailed', request => {
-      console.error(`[REQUEST FAILED] ${request.url()}: ${request.failure()?.errorText}`);
-    });
-    page.on('response', response => {
-      if (response.status() >= 400) {
-        console.error(`[HTTP ERROR] ${response.url()}: ${response.status()}`);
-      }
-    });
-  });
-
-  test("should render Midjourney mock and Sidepanel side-by-side and inject prompt", async ({ page }) => {
-    const screenshotsDir = path.join(__dirname, "../../tests/screenshots");
-    
-    try {
-      // 1. サンドボックス親ページを開く
-      console.log("Navigating to sandbox page...");
-      await page.goto("/tests/sandbox/index.html");
-
-      // 2. 左側 iframe (Midjourney) と 右側 iframe (サイドパネル) の取得
-      const mjFrame = page.frameLocator("#midjourney-frame");
-      const spFrame = page.frameLocator("#sidepanel-frame");
-
-      // 3. Midjourney モック内のテキストエリアが表示されるのを待つ
-      console.log("Waiting for Midjourney mock inputs...");
-      const textarea = mjFrame.locator('textarea, [role="textbox"], [data-testid="prompt-input"], [aria-label*="prompt"]').first();
-      await expect(textarea).toBeVisible({ timeout: 15000 });
-
-      // 4. サイドパネルのウェルカムダイアログの「スキップ」ボタンがあればクリック
-      console.log("Checking for welcome dialog...");
-      const skipButton = spFrame.locator("text=スキップ");
-      if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-        console.log("Welcome dialog detected, clicking skip...");
-        await skipButton.click();
-      }
-
-      // 5. HandBarの2枚目のカード（anime slot template）を削除し、1枚目のカードだけにする
-      console.log("Removing mock-card-2 from Workbench...");
-      const secondCard = spFrame.locator("#handbar-root .cursor-pointer").nth(1);
-      await expect(secondCard).toBeVisible({ timeout: 10000 });
-      await secondCard.hover();
-      const deleteBtn = secondCard.locator("button").last();
-      await deleteBtn.click();
-
-      // 6. サイドパネルの「Workbench」タブへ切り替え
-      console.log("Switching to Workbench tab in Sidepanel...");
-      const workbenchTabButton = spFrame.locator("button:has-text('Workbench')");
-      await expect(workbenchTabButton).toBeVisible({ timeout: 10000 });
-      await workbenchTabButton.click();
-
-      // 7. 「Workbench」の入力エリアにテキストを入力してプロンプトを追加
-      console.log("Adding prompt segments in Sidepanel...");
-      const promptInput = spFrame.locator("input.bg-transparent").first();
-      await expect(promptInput).toBeVisible({ timeout: 10000 });
-      await promptInput.fill("a cyberpunk ninja cat");
-      await promptInput.press("Enter");
-
-      // 8. 「Try on Midjourney」ボタンをクリックして注入を実行
-      console.log("Clicking 'Try on Midjourney' button...");
-      const injectButton = spFrame.locator("button:has-text('Try on Midjourney')");
-      await expect(injectButton).toBeVisible();
-      await injectButton.click();
-
-      // 9. Midjourney側のテキストエリアにプロンプトが正しく注入されたか検証
-      console.log("Verifying prompt injection in Midjourney mock...");
-      await expect(textarea).toHaveValue(
-        "neon-lit cyberpunk aesthetic, a cyberpunk ninja cat --sref https://midjourney.com/mock-sref",
-        { timeout: 10000 }
-      );
-      console.log("Prompt injection successfully verified!");
-
-      // 10. 視覚確認用のスクリーンショットを保存
-      await page.screenshot({
-        path: path.join(screenshotsDir, "side-by-side-success.png"),
-      });
-      console.log(`Success screenshot saved to tests/screenshots/side-by-side-success.png`);
-    } catch (error) {
-      console.error("Test failed, taking failure screenshot...");
-      try {
-        await page.screenshot({
-          path: path.join(screenshotsDir, "side-by-side-failure.png"),
-        });
-      } catch (err) {
-        console.error("Failed to capture failure screenshot:", err);
-      }
-      throw error;
-    }
-  });
-
-  test("should allow dragging a mock Midjourney image card into Sidepanel and saving it to History", async ({ page }) => {
-    const screenshotsDir = path.join(__dirname, "../../tests/screenshots");
-
-    try {
-      // 1. サンドボックス親ページを開く
-      console.log("Navigating to sandbox page for drag-and-drop test...");
-      await page.goto("/tests/sandbox/index.html");
-
-      const mjFrame = page.frameLocator("#midjourney-frame");
-      const spFrame = page.frameLocator("#sidepanel-frame");
-
-      // 2. Midjourney 内の最初のエレメント（超高層ビル画像）が表示されるのを待つ
-      console.log("Waiting for Midjourney mock images to render...");
-      const mockImage = mjFrame.locator("#pageScroll img").first();
-      await expect(mockImage).toBeVisible({ timeout: 15000 });
-
-      // 3. ウェルカムダイアログの「スキップ」ボタンがあればクリック
-      const skipButton = spFrame.locator("text=スキップ");
-      if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-        await skipButton.click();
-      }
-
-      // 4. ドラッグ元（画像）とドロップ先（サイドパネルの全体エリア）の特定
-      const dropZone = spFrame.locator(".h-full.relative.overflow-hidden");
-      await expect(dropZone).toBeVisible({ timeout: 10000 });
-
-      console.log("Dragging mock image from Midjourney frame to Sidepanel...");
-      
-      // クロスiframeのネイティブドラッグ＆ドロップを試行する
-      await mockImage.dragTo(dropZone, {
-        force: true
-      });
-
-      // 5. 万が一Playwrightのiframe間ドラッグがブラウザ仕様で不安定な場合のフォールバック:
-      // ドラッグのイベント発生が正しく仲介されていれば、サイドパネルに通知が表示される。
-      // もしドラッグが成功しなかった場合は、プログラム的にdropイベントを発火させて処理フローを補完・検証する。
-      const notification = spFrame.locator("text=New History Item Added!");
-      let isVisible = false;
-      try {
-        await expect(notification).toBeVisible({ timeout: 6000 });
-        isVisible = true;
-      } catch (e) {
-        isVisible = false;
-      }
-      
-      if (!isVisible) {
-        console.log("Fallback: Dispatching programmatic drop event to guarantee coverage...");
-        // 画像のメタデータを抽出して手動でイベントをディスパッチ
-        const imgData = {
-          id: "340ae0f9-c2f0-459c-ad45-95f7249049e8",
-          fullCommand: "超高層ビルを見上げた景色, 観葉植物, noon, skyscraper --ar 16:9 --sref 2496378872 3886212479 --stylize 200 --profile buibsja",
-          imageUrl: "./index_files/0_0_640_N.webp",
-          timestamp: Date.now()
-        };
-
-        await dropZone.evaluate(async (element, item) => {
-          const dataTransfer = new DataTransfer();
-          dataTransfer.setData("application/json", JSON.stringify(item));
-          
-          const dragOverEvent = new DragEvent("dragover", {
-            bubbles: true,
-            cancelable: true,
-            dataTransfer
-          });
-          element.dispatchEvent(dragOverEvent);
-
-          const dropEvent = new DragEvent("drop", {
-            bubbles: true,
-            cancelable: true,
-            dataTransfer
-          });
-          element.dispatchEvent(dropEvent);
-        }, imgData);
-      }
-
-      // 6. 「New History Item Added!」の通知が表示されることを確認
-      console.log("Checking drop success notification...");
-      await expect(notification).toBeVisible({ timeout: 10000 });
-
-      // 7. 「History」タブに切り替え
-      console.log("Switching to History tab...");
-      const historyTabButton = spFrame.locator("button:has-text('History')");
-      await expect(historyTabButton).toBeVisible();
-      await historyTabButton.click();
-
-      // 8. 注入したヒストリーアイテムが一覧に表示されることを検証
-      console.log("Verifying history list contains the dropped item...");
-      const historyItem = spFrame.locator("text=超高層ビルを見上げた景色").first();
-      await expect(historyItem).toBeVisible({ timeout: 10000 });
-
-      console.log("Drag-and-drop E2E test passed successfully!");
-      
-      // スクリーンショット保存
-      await page.screenshot({
-        path: path.join(screenshotsDir, "drag-and-drop-success.png"),
-      });
-    } catch (error) {
-      console.error("Drag-and-drop test failed, capturing screenshot...");
-      await page.screenshot({
-        path: path.join(screenshotsDir, "drag-and-drop-failure.png"),
-      });
-      throw error;
-    }
-  });
-
-  test("should correctly extract image prompts and image srefs from mock DOM", async ({ page }) => {
-    console.log("Navigating to sandbox page for extraction test...");
-    await page.goto("/tests/sandbox/index.html");
-
-    const mjFrame = page.frameLocator("#midjourney-frame");
-
-    // Wait for Midjourney mock images to render
-    console.log("Waiting for Midjourney mock images to render...");
-    const mockImage = mjFrame.locator("#pageScroll img").first();
-    await expect(mockImage).toBeVisible({ timeout: 15000 });
-
-    // Evaluate extractor inside midjourney-frame context
-    console.log("Evaluating WebDataExtractor on the image prompt row...");
-    const result = await mjFrame.locator("body").evaluate(() => {
-      // Find the first post image
-      const img = document.querySelector("#pageScroll img") as HTMLImageElement;
-      if (!img) return null;
-      
-      // Get the exposed extractor from window
-      const extractor = (window as any)._extractor;
-      if (!extractor) return { error: "Extractor not found on window" };
-
-      // Call extract and return the result
-      return extractor.extract(img);
-    });
-
-    console.log("Extracted result:", result);
-    expect(result).not.toBeNull();
-    expect(result.id).toBe("c8e0e10b-9a6f-4bd9-ab6c-d4252c60febb");
-    expect(result.fullCommand).toContain("https://s.mj.run/fOawY_NXKRY");
-    expect(result.fullCommand).toContain("超高層ビルを見上げた景色");
-    expect(result.fullCommand).toContain("カードが大量に重ねられて輝いている");
-    console.log("Extraction test passed successfully!");
-  });
-
-  test("should correctly extract Job ID, prompt, and image source from pattern2.html sandbox variant", async ({ page }) => {
-    console.log("Navigating to sandbox page with pattern2 variant...");
-    await page.goto("/tests/sandbox/index.html?variant=pattern2.html");
-
-    const mjFrame = page.frameLocator("#midjourney-frame");
-
-    // Wait for Midjourney mock images to render
-    console.log("Waiting for Midjourney mock images to render...");
-    const mockImage = mjFrame.locator("img[src*='0_1.jpeg']");
-    await expect(mockImage).toBeVisible({ timeout: 15000 });
-
-    // Evaluate extractor inside midjourney-frame context
-    console.log("Evaluating WebDataExtractor on the pattern2 image...");
-    const result = await mjFrame.locator("body").evaluate(() => {
-      const img = document.querySelector("img[src*='0_1.jpeg']") as HTMLImageElement;
-      if (!img) return null;
-      
-      const extractor = (window as any)._extractor;
-      if (!extractor) return { error: "Extractor not found on window" };
-
-      return extractor.extract(img);
-    });
-
-    console.log("Extracted result from pattern2:", result);
-    expect(result).not.toBeNull();
-    expect(result.id).toBe("100cc076-ef20-46b4-8aeb-f7c294169800");
-    expect(result.fullCommand).toContain("俊足の怪人");
-    expect(result.fullCommand).toContain("聖歌隊の行進");
-    expect(result.imageUrl).toContain("0_1.jpeg");
-    console.log("pattern2 variant extraction test passed successfully!");
-  });
-
-  test("should render slot variables and handle pin to hand in Workbench", async ({ page }) => {
-    console.log("Navigating to sandbox page for Workbench slot variable test...");
-    await page.goto("/tests/sandbox/index.html");
-
-    const spFrame = page.frameLocator("#sidepanel-frame");
-
-    // 1. ウェルカムダイアログの「スキップ」ボタンがあればクリック
-    const skipButton = spFrame.locator("text=スキップ");
-    if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-      await skipButton.click();
-    }
-
-    // 2. HandBarからanime slot templateをクリックしてWorkbenchに追加
-    const mockCardInHand = spFrame.locator("#handbar-root .cursor-pointer").nth(1); // anime slot template
-    await expect(mockCardInHand).toBeVisible({ timeout: 10000 });
-    await page.waitForTimeout(1000);
-    await mockCardInHand.click({ force: true });
-
-    // 3. Workbenchタブへ切り替え
-    const workbenchTabButton = spFrame.locator("button:has-text('Workbench')");
-    await expect(workbenchTabButton).toBeVisible({ timeout: 10000 });
-    await workbenchTabButton.click();
-
-    // 4. Slot Variablesセクションが表示されるのを確認
-    const slotSectionHeader = spFrame.locator("text=Slot Variables");
-    await expect(slotSectionHeader).toBeVisible({ timeout: 10000 });
-
-    // 5. Inputに値を入力して、Pin to Handボタンをクリック
-    const slotInput = spFrame.locator("input[placeholder='girl']");
-    await expect(slotInput).toBeVisible();
-    await slotInput.fill("samurai cat");
-    
-    // Send to Workbench
-    const sendBtn = spFrame.locator("button[title='Send to Workbench']").first();
-    await expect(sendBtn).toBeVisible();
-    await sendBtn.click();
-
-    // 6. 新しいカードがHandBarにピン留めされたかを確認
-    const newCardImage = spFrame.locator("#handbar-root img[alt='samurai cat']");
-    await expect(newCardImage).toBeVisible({ timeout: 10000 });
-  });
-
-  test("should open Merge Card Stack modal and execute merge in Workbench", async ({ page }) => {
-    console.log("Navigating to sandbox page for Workbench merge test...");
-    await page.goto("/tests/sandbox/index.html");
-
-    const spFrame = page.frameLocator("#sidepanel-frame");
-
-    // 1. スキップ
-    const skipButton = spFrame.locator("text=スキップ");
-    if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-      await skipButton.click();
-    }
-
-    // 2. 初期状態で両方のカードがすでにWorkbench（手札）に入っていることを確認
-    const cardCount = spFrame.locator("#handbar-root .cursor-pointer");
-    await expect(cardCount).toHaveCount(2);
-
-    // 3. Workbenchタブへ切り替え
-    const workbenchTabButton = spFrame.locator("button:has-text('Workbench')");
-    await workbenchTabButton.click();
-
-    // 4. Merge Stackボタンをクリックしてモーダルを開く
-    const mergeStackBtn = spFrame.locator("button:has-text('Merge Stack')").first();
-    await expect(mergeStackBtn).toBeVisible();
-    await mergeStackBtn.click();
-
-    // 5. モーダルの表示を確認
-    const modalTitle = spFrame.locator("h3:has-text('Merge Card Stack')");
-    await expect(modalTitle).toBeVisible();
-
-    // 6. モーダル内のMerge Stackを実行
-    const executeMergeBtn = spFrame.locator("button:has-text('Merge Stack')").nth(1);
-    await expect(executeMergeBtn).toBeVisible();
-    await executeMergeBtn.click();
-
-    // 7. モーダルが閉じたことを確認
-    await expect(modalTitle).not.toBeVisible({ timeout: 10000 });
-
-    // 材料カードが消費され、ベースカード1枚だけが残っていることを確認
-    const remainingCards = spFrame.locator("#handbar-root .cursor-pointer");
-    await expect(remainingCards).toHaveCount(1);
-  });
-
-  test("should allow managing tags in CardDetailView", async ({ page }) => {
-    console.log("Navigating to sandbox page for CardDetailView test...");
-    await page.goto("/tests/sandbox/index.html");
-
-    const spFrame = page.frameLocator("#sidepanel-frame");
-
-    // 1. スキップ
-    const skipButton = spFrame.locator("text=スキップ");
-    if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-      await skipButton.click();
-    }
-
-    // 2. Libraryタブに切り替え
-    const libraryTabButton = spFrame.locator("button:has-text('Library')");
-    await expect(libraryTabButton).toBeVisible();
-    await libraryTabButton.click();
-
-    // 3. カードの編集ボタンをクリックして詳細ビューを開く
-    const editBtn = spFrame.locator("[data-testid='edit-card-button']").first();
-    await expect(editBtn).toBeVisible({ timeout: 10000 });
-    await editBtn.click();
-
-    // 4. Card Detail Viewの表示確認
-    const detailTitle = spFrame.locator("h2:has-text('Card Details')");
-    await expect(detailTitle).toBeVisible();
-
-    // 5. タグの追加
-    const tagInput = spFrame.locator("input[placeholder='Add new tag...']");
-    await expect(tagInput).toBeVisible();
-    await tagInput.fill("e2e-tag");
-    const addTagBtn = spFrame.locator("button:has-text('Add')");
-    await addTagBtn.click();
-
-    // 6. タグが表示されたことを確認
-    const newTagChip = spFrame.locator("text=e2e-tag");
-    await expect(newTagChip).toBeVisible();
-
-    // 7. 保存して閉じる
-    const saveBtn = spFrame.locator("button:has-text('Save')");
-    await saveBtn.click();
-
-    // 8. 詳細ビューが閉じたことを確認
-    await expect(detailTitle).not.toBeVisible({ timeout: 5000 });
-  });
-
-  test("should load history items reactively and load more items via infinite scroll", async ({ page }) => {
-    const screenshotsDir = path.join(__dirname, "../../tests/screenshots");
-    
-    console.log("Navigating to sandbox page for infinite scroll test...");
-    await page.goto("/tests/sandbox/index.html");
-    
-    const spFrame = page.frameLocator("#sidepanel-frame");
-    
-    // スキップ
-    const skipButton = spFrame.locator("text=スキップ");
-    if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-      await skipButton.click();
-    }
-    
-    // E2Eテスト用のダミー履歴データを大量に挿入する (70件)
-    console.log("Seeding 70 mock history items into IndexedDB...");
-    await spFrame.locator("body").evaluate(async () => {
-      const database = (window as any).db;
-      if (!database) {
-        throw new Error("Database instance not found on window");
-      }
-      await database.historyItems.clear();
-      
-      const mockItems = Array.from({ length: 70 }, (_, i) => ({
-        id: `mock-history-${i}`,
-        fullCommand: `Cyberpunk prompt number ${i} --ar 16:9`,
-        imageUrl: `./index_files/0_0_640_N.webp`,
-        timestamp: Date.now() - i * 1000
-      }));
-      await database.historyItems.bulkAdd(mockItems);
-    });
-
-    // Historyタブへ切り替え
-    console.log("Switching to History tab...");
-    const historyTabButton = spFrame.locator("button:has-text('History')");
-    await expect(historyTabButton).toBeVisible();
-    await historyTabButton.click();
-
-    // 初期状態で50件のみ表示されていることを確認
-    console.log("Verifying initial 50 items are displayed...");
-    const mintButtons = spFrame.locator("button:has-text('Mint Card')");
-    await expect(mintButtons).toHaveCount(50, { timeout: 10000 });
-
-    // 50番目のカードまでスクロールする
-    console.log("Scrolling to the bottom...");
-    const lastCard = mintButtons.nth(49);
-    await lastCard.scrollIntoViewIfNeeded();
-
-    // センチネルが作動して追加の20件がロードされ、合計70件になることを確認
-    console.log("Verifying additional items loaded...");
-    await expect(mintButtons).toHaveCount(70, { timeout: 10000 });
-    
-    // スクリーンショット保存
-    await page.screenshot({
-      path: path.join(screenshotsDir, "infinite-scroll-success.png"),
-    });
-    console.log("Infinite scroll E2E test passed successfully!");
-  });
-
-  test("should allow creating, editing, and deleting a custom category in Library tab", async ({ page }) => {
-    const screenshotsDir = path.join(__dirname, "../../tests/screenshots");
-
-    page.on('console', msg => {
-      console.log(`[CATEGORY TEST CONSOLE] ${msg.type()}: ${msg.text()}`);
-    });
-    page.on('pageerror', err => {
-      console.error(`[CATEGORY TEST ERROR] ${err.message}\n${err.stack}`);
-    });
-
-    try {
-      console.log("1. Navigating to sandbox page...");
-      await page.goto("/tests/sandbox/index.html");
-
-      const spFrame = page.frameLocator("#sidepanel-frame");
-
-      // 1. ウェルカムダイアログの「スキップ」ボタンがあればクリック
-      console.log("2. Checking for welcome dialog...");
-      const skipButton = spFrame.locator("text=スキップ");
-      if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-        console.log("Clicking skip button...");
-        await skipButton.click();
-        await page.waitForTimeout(500);
-      }
-
-      // 2. Libraryタブに切り替え
-      console.log("3. Switching to Library tab...");
-      const libraryTabButton = spFrame.locator("button:has-text('Library')");
-      await expect(libraryTabButton).toBeVisible({ timeout: 15000 });
-      await libraryTabButton.click();
-      await page.waitForTimeout(1000); // Wait for Dexie queries to load style cards and categories
-
-      // 3. タグボタン（Manage Categories）をクリックしてモーダルを開く
-      console.log("4. Opening category modal...");
-      const addCategoryBtn = spFrame.locator("button[title='Manage Categories']");
-      await expect(addCategoryBtn).toBeVisible({ timeout: 15000 });
-      await addCategoryBtn.click();
-      await page.waitForTimeout(500);
-
-      // 4. モーダルの入力フィールドに入力してカスタムカテゴリを作成
-      console.log("5. Creating new category...");
-      const nameInput = spFrame.locator("input[placeholder='e.g. Cyberpunk, Retro']");
-      await expect(nameInput).toBeVisible({ timeout: 10000 });
-      await nameInput.fill("E2E Category");
-
-      const emojiInput = spFrame.locator("input[placeholder='e.g. 🎨, 🛸']");
-      await emojiInput.fill("👾");
-
-      const submitBtn = spFrame.locator("button:has-text('Create Category')");
-      await submitBtn.click();
-      await page.waitForTimeout(1000); // Wait for category to be added to IndexedDB and UI to update
-
-      // 5. 新しいカテゴリフィルターボタンがLibraryタブに追加されたか確認
-      console.log("6. Verifying category filter button added...");
-      const newCategoryFilterBtn = spFrame.locator("button:has-text('E2E Category')");
-      await expect(newCategoryFilterBtn).toBeVisible({ timeout: 15000 });
-
-      // 6. 再び "+" ボタンをクリックして管理モーダルを開く
-      console.log("7. Re-opening category modal...");
-      await addCategoryBtn.click();
-      await page.waitForTimeout(500);
-
-      // 7. 「Manage Categories」タブに切り替え
-      console.log("8. Switching to Manage Categories tab...");
-      const manageTabBtn = spFrame.locator("button:has-text('Manage Categories')");
-      await expect(manageTabBtn).toBeVisible({ timeout: 10000 });
-      await manageTabBtn.click();
-      await page.waitForTimeout(1000); // Wait for categories list to load
-
-      // 8. 作成したカスタムカテゴリが表示されているか確認
-      console.log("9. Verifying category row exists...");
-      const categoryRow = spFrame.locator("p", { hasText: "E2E Category" }).first();
-      await expect(categoryRow).toBeVisible({ timeout: 10000 });
-
-      // 9. 編集ボタンをクリックして編集画面へ
-      console.log("10. Clicking Edit Category button...");
-      const editBtn = spFrame.getByRole("button", { name: "Edit Category" }).first();
-      await expect(editBtn).toBeVisible({ timeout: 10000 });
-      await editBtn.click();
-      await page.waitForTimeout(500);
-
-      // 10. 名前の変更と保存
-      console.log("11. Changing name and saving...");
-      await nameInput.fill("E2E Category Edited");
-      const saveBtn = spFrame.locator("button:has-text('Save Changes')");
-      await saveBtn.click();
-      await page.waitForTimeout(1000); // Wait for DB write and UI update
-
-      // 11. 変更したカテゴリ名がフィルター行に表示されているか確認
-      console.log("12. Verifying edited category filter button...");
-      const editedCategoryFilterBtn = spFrame.locator("button:has-text('E2E Category Edited')");
-      await expect(editedCategoryFilterBtn).toBeVisible({ timeout: 15000 });
-
-      // 12. 削除する (モーダルは既に開いていてManage tabになっているので、そのまま削除を実行できる)
-      // 削除用 confirm をモックする
-      console.log("13. Registering dialog mock...");
-      page.once("dialog", async (dialog) => {
-        expect(dialog.message()).toContain("Are you sure you want to delete");
-        await dialog.accept();
-      });
-
-      console.log("14. Clicking Delete Category button...");
-      const deleteBtn = spFrame.getByRole("button", { name: "Delete Category" }).first();
-      await expect(deleteBtn).toBeVisible({ timeout: 10000 });
-      await deleteBtn.click();
-      await page.waitForTimeout(1000);
-
-      // 13. モーダルを閉じる
-      console.log("15. Closing category modal...");
-      const closeBtn = spFrame.getByRole("button", { name: "Close" }).first();
-      await expect(closeBtn).toBeVisible({ timeout: 10000 });
-      await closeBtn.click();
-      await page.waitForTimeout(500);
-
-      // 14. フィルター行からカテゴリが消えたことを確認
-      console.log("16. Verifying category filter is deleted...");
-      await expect(editedCategoryFilterBtn).not.toBeVisible({ timeout: 15000 });
-
-      console.log("Custom category E2E test passed successfully!");
-
-      // スクリーンショット保存
-      await page.screenshot({
-        path: path.join(screenshotsDir, "custom-category-success.png"),
-      });
-    } catch (error) {
-      console.error("Custom category E2E test failed, capturing failure screenshot...");
-      await page.screenshot({
-        path: path.join(screenshotsDir, "custom-category-failure.png"),
-      });
-      throw error;
-    }
-  });
-
-  test("should allow minting a new card from History (Scenario 2)", async ({ page }) => {
-    const screenshotsDir = path.join(__dirname, "../../tests/screenshots");
-    console.log("Navigating to sandbox page for Minting E2E test...");
-    await page.goto("/tests/sandbox/index.html");
-
-    const spFrame = page.frameLocator("#sidepanel-frame");
-
-    // 1. Skip welcome dialog
-    const skipButton = spFrame.locator("text=スキップ");
-    if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-      await skipButton.click();
-    }
-
-    // 2. Clear and seed history items to guarantee the item is available
-    await spFrame.locator("body").evaluate(async () => {
-      const database = (window as any).db;
-      await database.historyItems.clear();
-      await database.historyItems.add({
-        id: "mock-history-item-to-mint",
-        fullCommand: "a beautiful cyberpunk warrior --ar 16:9 --sref https://example.com/sref1 --p p-code",
-        imageUrl: "./index_files/0_0_640_N.webp",
-        timestamp: Date.now()
-      });
-    });
-
-    // 3. Switch to History tab
-    const historyTabButton = spFrame.locator("button:has-text('History')");
-    await expect(historyTabButton).toBeVisible();
-    await historyTabButton.click();
-
-    // 4. Click "Mint Card" button
-    const mintCardBtn = spFrame.locator("button:has-text('Mint Card')").first();
-    await expect(mintCardBtn).toBeVisible({ timeout: 10000 });
-    await mintCardBtn.click();
-
-    // 5. Verify minting view container is visible
-    const mintingView = spFrame.locator("[data-testid='minting-view-container']");
-    await expect(mintingView).toBeVisible({ timeout: 10000 });
-
-    // 6. Enter Custom Name
-    const nameInput = spFrame.locator("input[placeholder='Add details...']");
-    await nameInput.fill("Warrior Note");
-
-    // 7. Select rarity "Epic"
-    const rarityBtn = spFrame.locator("button:has-text('Epic')");
-    await expect(rarityBtn).toBeVisible();
-    await rarityBtn.click();
-
-    // 8. Click Save Card
-    const saveCardBtn = spFrame.locator("button:has-text('Save Card')");
-    await expect(saveCardBtn).toBeVisible();
-    await saveCardBtn.click();
-
-    // 9. Verify minting view is closed
-    await expect(mintingView).not.toBeVisible({ timeout: 10000 });
-
-    // 10. Switch to Library tab and verify the card is there
-    const libraryTabButton = spFrame.locator("button:has-text('Library')");
-    await libraryTabButton.click();
-    await page.waitForTimeout(1000); // wait for DB query
-    const cardTitle = spFrame.locator("text=Warrior Note").first();
-    await expect(cardTitle).toBeVisible({ timeout: 10000 });
-
-    await page.screenshot({
-      path: path.join(screenshotsDir, "mint-success.png"),
-    });
-  });
-
-  test("should associate dropped history image of same job ID and allow editing thumbnails (Scenario 3)", async ({ page }) => {
-    const screenshotsDir = path.join(__dirname, "../../tests/screenshots");
-    console.log("Navigating to sandbox page for same job image association E2E test...");
-    await page.goto("/tests/sandbox/index.html");
-
-    const spFrame = page.frameLocator("#sidepanel-frame");
-
-    // 1. Skip welcome dialog
-    const skipButton = spFrame.locator("text=スキップ");
-    if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-      await skipButton.click();
-    }
-
-    // 2. Seed a Style Card in DB
-    await spFrame.locator("body").evaluate(async () => {
-      const database = (window as any).db;
-      await database.styleCards.clear();
-      await database.styleCards.add({
-        id: "test-card-same-job",
-        name: "Same Job Test Card",
-        jobId: "job-id-12345",
-        promptSegments: [{ type: "text", value: "testing same job" }],
-        parameters: {},
-        masking: { isSrefHidden: false, isPHidden: false },
-        tier: "Common",
-        isFavorite: false,
-        isPinned: false,
-        usageCount: 0,
-        tags: [],
-        dominantColor: "#3b82f6",
-        thumbnailData: "data:image/svg+xml;utf8,<svg></svg>",
-        images: ["https://example.com/original-image.png"],
-        selectedThumbnails: ["https://example.com/original-image.png"],
-        associatedJobIds: []
-      });
-    });
-
-    // 3. Dispatch a drop event with the same job ID but a new image URL
-    console.log("Simulating drag-and-drop of an image from the same job ID...");
-    const dropZone = spFrame.locator(".h-full.relative.overflow-hidden");
-    const imgData = {
-      id: "job-id-12345", // Same job ID!
-      fullCommand: "testing same job --ar 16:9",
-      imageUrl: "https://example.com/new-associated-image.png", // Different image URL!
-      timestamp: Date.now()
-    };
-    await dropZone.evaluate(async (element, item) => {
-      const dataTransfer = new DataTransfer();
-      dataTransfer.setData("application/json", JSON.stringify(item));
-      const dragOverEvent = new DragEvent("dragover", { bubbles: true, cancelable: true, dataTransfer });
-      element.dispatchEvent(dragOverEvent);
-      const dropEvent = new DragEvent("drop", { bubbles: true, cancelable: true, dataTransfer });
-      element.dispatchEvent(dropEvent);
-    }, imgData);
-
-    // 4. Verify notification toast shows card name updated
-    const notification = spFrame.locator("text=Same Job Test Card").first();
-    await expect(notification).toBeVisible({ timeout: 10000 });
-
-    // 5. Switch to Library tab
-    const libraryTabButton = spFrame.locator("button:has-text('Library')");
-    await libraryTabButton.click();
-    await page.waitForTimeout(1000); // wait for DB query
-
-    // 6. Click edit button on the card
-    const editBtn = spFrame.locator("[data-testid='edit-card-button']").first();
-    await expect(editBtn).toBeVisible();
-    await editBtn.click();
-
-    // 7. Verify CardDetailView is open
-    const detailTitle = spFrame.locator("h2:has-text('Card Details')");
-    await expect(detailTitle).toBeVisible();
-
-    // 8. Verify Associated Images count is 2
-    const associatedTitle = spFrame.locator("h3:has-text('Associated Images (2)')");
-    await expect(associatedTitle).toBeVisible({ timeout: 10000 });
-
-    // 9. Click on the 2nd image (newly added) to select it as thumbnail
-    const secondImageCell = spFrame.locator("img[alt='Card Image 2']");
-    await expect(secondImageCell).toBeVisible();
-    await secondImageCell.click();
-
-    // 10. Verify selected thumbnails count is 2 / 4
-    const selectedBadge = spFrame.locator("text=Selected: 2 / 4");
-    await expect(selectedBadge).toBeVisible();
-
-    // 11. Click Save
-    const saveBtn = spFrame.locator("button:has-text('Save')");
-    await saveBtn.click();
-
-    // 12. Verify CardDetailView is closed
-    await expect(detailTitle).not.toBeVisible({ timeout: 10000 });
-
-    // 13. Verify in DB that selectedThumbnails contains the new URL
-    const selectedThumbnailsInDb = await spFrame.locator("body").evaluate(async () => {
-      const database = (window as any).db;
-      const card = await database.getCard("test-card-same-job");
-      return card?.selectedThumbnails || [];
-    });
-    expect(selectedThumbnailsInDb).toContain("https://example.com/new-associated-image.png");
-
-    await page.screenshot({
-      path: path.join(screenshotsDir, "same-job-success.png"),
-    });
-  });
-
-  test("should filter card list by tag search in Library (Scenario 4)", async ({ page }) => {
-    const screenshotsDir = path.join(__dirname, "../../tests/screenshots");
-    console.log("Navigating to sandbox page for tag search E2E test...");
-    await page.goto("/tests/sandbox/index.html");
-
-    const spFrame = page.frameLocator("#sidepanel-frame");
-
-    // 1. Skip welcome dialog
-    const skipButton = spFrame.locator("text=スキップ");
-    if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-      await skipButton.click();
-    }
-
-    // 2. Seed two cards with different tags
-    await spFrame.locator("body").evaluate(async () => {
-      const database = (window as any).db;
-      await database.styleCards.clear();
-      await database.styleCards.bulkAdd([
-        {
-          id: "card-vintage",
-          name: "Vintage Card",
-          promptSegments: [{ type: "text", value: "vintage prompt" }],
-          parameters: {},
-          masking: {},
-          tier: "Common",
-          tags: ["vintage", "retro"],
-          thumbnailData: "data:image/svg+xml;utf8,<svg></svg>"
-        },
-        {
-          id: "card-futuristic",
-          name: "Futuristic Card",
-          promptSegments: [{ type: "text", value: "futuristic prompt" }],
-          parameters: {},
-          masking: {},
-          tier: "Common",
-          tags: ["cyberpunk", "neon"],
-          thumbnailData: "data:image/svg+xml;utf8,<svg></svg>"
-        }
-      ]);
-    });
-
-    // 3. Switch to Library tab
-    const libraryTabButton = spFrame.locator("button:has-text('Library')");
-    await libraryTabButton.click();
-    await page.waitForTimeout(1000); // wait for DB queries
-
-    // 4. Fill search field with "cyberpunk"
-    const searchInput = spFrame.locator("input[placeholder*='Search']").first();
-    await expect(searchInput).toBeVisible();
-    await searchInput.fill("cyberpunk");
-    await page.waitForTimeout(500); // wait for search filtering
-
-    // 5. Verify "Futuristic Card" is visible, "Vintage Card" is NOT visible
-    await expect(spFrame.locator("text=Futuristic Card")).toBeVisible();
-    await expect(spFrame.locator("text=Vintage Card")).not.toBeVisible();
-
-    // 6. Clear search, type "vintage"
-    await searchInput.fill("vintage");
-    await page.waitForTimeout(500); // wait for search filtering
-
-    // 7. Verify "Vintage Card" is visible, "Futuristic Card" is NOT visible
-    await expect(spFrame.locator("text=Vintage Card")).toBeVisible();
-    await expect(spFrame.locator("text=Futuristic Card")).not.toBeVisible();
-
-    await page.screenshot({
-      path: path.join(screenshotsDir, "search-success.png"),
-    });
-  });
-
-  test("should support Workbench-Single flow (Scenario 5)", async ({ page }) => {
-    const screenshotsDir = path.join(__dirname, "../../tests/screenshots");
-    console.log("Navigating to sandbox page for Workbench-Single E2E test...");
-    await page.goto("/tests/sandbox/index.html");
-
-    const mjFrame = page.frameLocator("#midjourney-frame");
-    const spFrame = page.frameLocator("#sidepanel-frame");
-
-    // 1. Skip welcome dialog
-    const skipButton = spFrame.locator("text=スキップ");
-    if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-      await skipButton.click();
-    }
-
-    // 2. Seed 3 cards, all pinned (so they are in the Hand)
-    await spFrame.locator("body").evaluate(async () => {
-      const database = (window as any).db;
-      await database.styleCards.clear();
-      await database.styleCards.bulkAdd([
-        {
-          id: "card-w1",
-          name: "Card W1",
-          promptSegments: [
-            { type: "text", value: "cyberpunk cat" },
-            { type: "slot", label: "Color", default: "blue" }
-          ],
-          parameters: { sref: ["https://example.com/sref1"] },
-          masking: { isSrefHidden: false, isPHidden: false },
-          tier: "Common",
-          isPinned: true,
-          dominantColor: "#3b82f6",
-          thumbnailData: "data:image/svg+xml;utf8,<svg></svg>"
-        },
-        {
-          id: "card-w2",
-          name: "Card W2",
-          promptSegments: [{ type: "text", value: "neon glow" }],
-          parameters: {},
-          masking: {},
-          tier: "Rare",
-          isPinned: true,
-          dominantColor: "#ec4899",
-          thumbnailData: "data:image/svg+xml;utf8,<svg></svg>"
-        },
-        {
-          id: "card-w3",
-          name: "Card W3",
-          promptSegments: [{ type: "text", value: "synthwave sun" }],
-          parameters: {},
-          masking: {},
-          tier: "Epic",
-          isPinned: true,
-          dominantColor: "#8b5cf6",
-          thumbnailData: "data:image/svg+xml;utf8,<svg></svg>"
-        }
-      ]);
-    });
-
-    // 3. Switch to Workbench tab
-    const workbenchTabButton = spFrame.locator("button:has-text('Workbench')");
-    await workbenchTabButton.click();
-    await page.waitForTimeout(1000); // wait for DB queries
-
-    // Unpin Card W2 and Card W3 using the X button on their thumbnails in the HandBar
-    console.log("Unpinning Card W2 and Card W3...");
-    const cardW2 = spFrame.locator("#handbar-root .cursor-pointer").nth(1);
-    await cardW2.hover();
-    await cardW2.locator("button").last().click();
-
-    const cardW3 = spFrame.locator("#handbar-root .cursor-pointer").nth(1);
-    await cardW3.hover();
-    await cardW3.locator("button").last().click();
-
-    // Now only 1 card (Card W1) is in the Hand/Workbench
-    const activeWorkbenchCards = spFrame.locator("#handbar-root .cursor-pointer");
-    await expect(activeWorkbenchCards).toHaveCount(1);
-
-    // 4. Fill in slot value for "Color"
-    const slotInput = spFrame.locator("input[placeholder='blue']");
-    await expect(slotInput).toBeVisible({ timeout: 10000 });
-    await slotInput.fill("purple");
-
-    // 5. Try on Midjourney (Inject)
-    const injectBtn = spFrame.locator("button:has-text('Try on Midjourney')");
-    await expect(injectBtn).toBeVisible();
-    await injectBtn.click();
-
-    // 6. Verify prompt in Midjourney mock textarea
-    const mjTextarea = mjFrame.locator('textarea, [role="textbox"], [data-testid="prompt-input"]').first();
-    await expect(mjTextarea).toHaveValue("cyberpunk cat, purple --sref https://example.com/sref1", { timeout: 10000 });
-
-    await page.screenshot({
-      path: path.join(screenshotsDir, "workbench-single-success.png"),
-    });
-  });
-
-  test("should support Workbench-Double flow (Scenario 6)", async ({ page }) => {
-    const screenshotsDir = path.join(__dirname, "../../tests/screenshots");
-    console.log("Navigating to sandbox page for Workbench-Double E2E test...");
-    await page.goto("/tests/sandbox/index.html");
-
-    const mjFrame = page.frameLocator("#midjourney-frame");
-    const spFrame = page.frameLocator("#sidepanel-frame");
-
-    // 1. Skip welcome dialog
-    const skipButton = spFrame.locator("text=スキップ");
-    if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-      await skipButton.click();
-    }
-
-    // 2. Seed 3 cards, all pinned (so they are in the Hand)
-    await spFrame.locator("body").evaluate(async () => {
-      const database = (window as any).db;
-      await database.styleCards.clear();
-      await database.styleCards.bulkAdd([
-        {
-          id: "card-w1",
-          name: "Card W1",
-          promptSegments: [
-            { type: "text", value: "cyberpunk cat" },
-            { type: "slot", label: "Color", default: "blue" }
-          ],
-          parameters: { sref: ["https://example.com/sref1"] },
-          masking: { isSrefHidden: false, isPHidden: false },
-          tier: "Common",
-          isPinned: true,
-          dominantColor: "#3b82f6",
-          thumbnailData: "data:image/svg+xml;utf8,<svg></svg>"
-        },
-        {
-          id: "card-w2",
-          name: "Card W2",
-          promptSegments: [{ type: "text", value: "neon glow" }],
-          parameters: { p: ["p-code-w2"] },
-          masking: { isSrefHidden: false, isPHidden: false },
-          tier: "Rare",
-          isPinned: true,
-          dominantColor: "#ec4899",
-          thumbnailData: "data:image/svg+xml;utf8,<svg></svg>"
-        },
-        {
-          id: "card-w3",
-          name: "Card W3",
-          promptSegments: [{ type: "text", value: "synthwave sun" }],
-          parameters: {},
-          masking: {},
-          tier: "Epic",
-          isPinned: true,
-          dominantColor: "#8b5cf6",
-          thumbnailData: "data:image/svg+xml;utf8,<svg></svg>"
-        }
-      ]);
-    });
-
-    // 3. Switch to Workbench tab
-    const workbenchTabButton = spFrame.locator("button:has-text('Workbench')");
-    await workbenchTabButton.click();
-    await page.waitForTimeout(1000); // wait for DB queries
-
-    // Workbench-Double: select two cards (W1 and W2). So unpin Card W3 (index 2).
-    console.log("Unpinning Card W3...");
-    const cardW3 = spFrame.locator("#handbar-root .cursor-pointer").nth(2);
-    await cardW3.hover();
-    await cardW3.locator("button").last().click();
-
-    // Now 2 cards are in the Hand/Workbench.
-    const activeWorkbenchCards = spFrame.locator("#handbar-root .cursor-pointer");
-    await expect(activeWorkbenchCards).toHaveCount(2);
-
-    // 4. Fill in slot value for "Color"
-    const slotInput = spFrame.locator("input[placeholder='blue']");
-    await expect(slotInput).toBeVisible({ timeout: 10000 });
-    await slotInput.fill("yellow");
-
-    // 5. Try on Midjourney (Inject)
-    const injectBtn = spFrame.locator("button:has-text('Try on Midjourney')");
-    await expect(injectBtn).toBeVisible();
-    await injectBtn.click();
-
-    // 6. Verify prompt in Midjourney mock textarea
-    const mjTextarea = mjFrame.locator('textarea, [role="textbox"], [data-testid="prompt-input"]').first();
-    await expect(mjTextarea).toHaveValue("cyberpunk cat, yellow, neon glow --sref https://example.com/sref1 --p p-code-w2", { timeout: 10000 });
-
-    await page.screenshot({
-      path: path.join(screenshotsDir, "workbench-double-success.png"),
-    });
-  });
-
-  test("should support Workbench-Triple flow (Scenario 7)", async ({ page }) => {
-    const screenshotsDir = path.join(__dirname, "../../tests/screenshots");
-    console.log("Navigating to sandbox page for Workbench-Triple E2E test...");
-    await page.goto("/tests/sandbox/index.html");
-
-    const spFrame = page.frameLocator("#sidepanel-frame");
-
-    // 1. Skip welcome dialog
-    const skipButton = spFrame.locator("text=スキップ");
-    if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-      await skipButton.click();
-    }
-
-    // 2. Seed 3 cards, all pinned (so they are in the Hand)
-    await spFrame.locator("body").evaluate(async () => {
-      const database = (window as any).db;
-      await database.styleCards.clear();
-      await database.styleCards.bulkAdd([
-        {
-          id: "card-w1",
-          name: "Card W1",
-          promptSegments: [{ type: "text", value: "cyberpunk cat" }],
-          parameters: { sref: ["https://example.com/sref1"] },
-          masking: {},
-          tier: "Common",
-          isPinned: true,
-          dominantColor: "#3b82f6",
-          thumbnailData: "data:image/svg+xml;utf8,<svg></svg>",
-          usageCount: 5
-        },
-        {
-          id: "card-w2",
-          name: "Card W2",
-          promptSegments: [{ type: "text", value: "neon glow" }],
-          parameters: {},
-          masking: {},
-          tier: "Rare",
-          isPinned: true,
-          dominantColor: "#ec4899",
-          thumbnailData: "data:image/svg+xml;utf8,<svg></svg>",
-          usageCount: 3
-        },
-        {
-          id: "card-w3",
-          name: "Card W3",
-          promptSegments: [{ type: "text", value: "synthwave sun" }],
-          parameters: {},
-          masking: {},
-          tier: "Epic",
-          isPinned: true,
-          dominantColor: "#8b5cf6",
-          thumbnailData: "data:image/svg+xml;utf8,<svg></svg>",
-          usageCount: 2
-        }
-      ]);
-    });
-
-    // 3. Switch to Workbench tab
-    const workbenchTabButton = spFrame.locator("button:has-text('Workbench')");
-    await workbenchTabButton.click();
-    await page.waitForTimeout(1000); // wait for DB queries
-
-    // 4. Click "Merge Stack" button in the HandBar
-    const mergeStackBtn = spFrame.locator("[data-testid='handbar-merge-btn']");
-    await expect(mergeStackBtn).toBeVisible({ timeout: 10000 });
-    await mergeStackBtn.click();
-
-    // 5. Verify Merge Modal is open
-    const modalTitle = spFrame.locator("h3:has-text('Merge Card Stack')");
-    await expect(modalTitle).toBeVisible();
-
-    // 6. Base Card W1 is selected by default (index 0).
-    // Let's verify that Card W2 is in the Material list and is set to "Consume" by default.
-    // And Card W3 is also in the Material list and is set to "Consume". We click on Card W3's "Consume" button to change it to "Keep".
-    console.log("Setting Card W3 integration to Keep...");
-    const cardW3Row = spFrame.locator(".space-y-2").last().locator("div", { has: spFrame.locator("p", { hasText: "Card W3" }) }).first();
-    const consumeBtn = cardW3Row.locator("button:has-text('Consume')");
-    await expect(consumeBtn).toBeVisible();
-    await consumeBtn.click();
-
-    // Verify it is now "Keep"
-    await expect(cardW3Row.locator("button:has-text('Keep')")).toBeVisible();
-
-    // 7. Click "Merge Stack" button inside the modal to execute merge
-    const modalExecuteBtn = spFrame.locator("button:has-text('Merge Stack')").nth(1);
-    await modalExecuteBtn.click();
-
-    // 8. Verify modal is closed
-    await expect(modalTitle).not.toBeVisible({ timeout: 10000 });
-    await page.waitForTimeout(1000); // wait for DB transaction
-
-    // 9. Verify in DB that W1 usage count is combined (5 + 3 = 8), W2 is soft-deleted, and W3 remains.
-    const results = await spFrame.locator("body").evaluate(async () => {
-      const database = (window as any).db;
-      const w1 = await database.getCard("card-w1");
-      const w2 = await database.getCard("card-w2");
-      const w3 = await database.getCard("card-w3");
-      return {
-        w1Exists: !!w1,
-        w1Usage: w1?.usageCount,
-        w2Exists: !!w2,
-        w3Exists: !!w3
-      };
-    });
-
-    console.log("Merge verification results:", results);
-    expect(results.w1Exists).toBe(true);
-    expect(results.w1Usage).toBe(8);
-    expect(results.w2Exists).toBe(false);
-    expect(results.w3Exists).toBe(true);
-
-    await page.screenshot({
-      path: path.join(screenshotsDir, "workbench-triple-success.png"),
-    });
-  });
-
-  test("should allow selecting restore mode and verify replace/merge behavior", async ({ page }) => {
-    const screenshotsDir = path.join(__dirname, "../../tests/screenshots");
-
-    // Mock Google Drive API response
-    await page.route("https://www.googleapis.com/drive/v3/files*", async (route) => {
-      await route.fulfill({
-        status: 200,
-        contentType: "application/json",
-        body: JSON.stringify({
-          files: [
-            {
-              id: "mock-file-123",
-              modifiedTime: "2026-06-03 12:00:00",
-              size: "153600" // 150.0 KB
-            }
-          ]
-        })
-      });
-    });
-
-    console.log("Navigating to sandbox page for Restore Mode E2E test...");
-    await page.goto("/tests/sandbox/index.html");
-
-    const spFrame = page.frameLocator("#sidepanel-frame");
-
-    // 1. Skip welcome dialog
-    const skipButton = spFrame.locator("text=スキップ");
-    if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-      await skipButton.click();
-    }
-
-    // 2. Switch to Settings tab
-    const settingsTabButton = spFrame.locator("nav button:has-text('Settings')");
-    await expect(settingsTabButton).toBeVisible();
-    await settingsTabButton.click();
-
-    // 3. Verify Sync and Force Recovery UI is visible
-    const syncBtn = spFrame.locator("#google-drive-sync-btn");
-    const forceRecoveryBtn = spFrame.locator("#force-recovery-btn");
-    await expect(syncBtn).toBeVisible({ timeout: 10000 });
-    await expect(forceRecoveryBtn).toBeVisible();
-
-    // Enable Google Drive synchronization
-    console.log("Enabling Google Drive sync...");
-    const toggleBtn = spFrame.locator("#google-drive-toggle-btn");
-    await expect(toggleBtn).toBeVisible();
-    await toggleBtn.click();
-
-    // Wait for the cloud backup preview to appear
-    console.log("Waiting for cloud backup preview...");
-    const previewEl = spFrame.locator("text=150.0 KB").first();
-    await expect(previewEl).toBeVisible({ timeout: 10000 });
-
-    // Scroll to the Force Recovery button to bring the Danger Zone preview into view
-    await forceRecoveryBtn.scrollIntoViewIfNeeded();
-
-    // 4. Capture screenshot of Settings tab with new UI
-    await page.screenshot({
-      path: path.join(screenshotsDir, "restore-mode-ui.png"),
-    });
-
-    // 5. Test sync/merge and soft delete behavior via page evaluation (Dexie verification)
-    const results = await spFrame.locator("body").evaluate(async () => {
-      const database = (window as any).db;
-      
-      const resetData = async () => {
-        await database.styleCards.clear();
-        await database.categories.clear();
-        await database.userSettings.clear();
-        await database.historyItems.clear();
-      };
-
-      // Set initial local state
-      await resetData();
-      await database.styleCards.bulkAdd([
-        {
-          id: "card-1",
-          name: "Local Card 1 (Old)",
-          createdAt: 1000,
-          updatedAt: 1000,
-          promptSegments: [],
-          parameters: {},
-          masking: {},
-          tier: "Common",
-          tags: [],
-          dominantColor: "#000",
-          thumbnailData: ""
-        },
-        {
-          id: "card-2",
-          name: "Local Only Card",
-          createdAt: 1000,
-          updatedAt: 1000,
-          promptSegments: [],
-          parameters: {},
-          masking: {},
-          tier: "Common",
-          tags: [],
-          dominantColor: "#000",
-          thumbnailData: ""
-        }
-      ]);
-
-      const backupPayload = {
-        styleCards: [
-          {
-            id: "card-1",
-            name: "Backup Card 1 (New)",
-            createdAt: 1000,
-            updatedAt: 2000, // Newer!
-            promptSegments: [],
-            parameters: {},
-            masking: {},
-            tier: "Common",
-            tags: [],
-            dominantColor: "#000",
-            thumbnailData: ""
-          },
-          {
-            id: "card-3",
-            name: "Backup Only Card",
-            createdAt: 1500,
-            updatedAt: 1500,
-            promptSegments: [],
-            parameters: {},
-            masking: {},
-            tier: "Common",
-            tags: [],
-            dominantColor: "#000",
-            thumbnailData: ""
-          }
-        ],
-        categories: [],
-        userSettings: [],
-        historyItems: []
-      };
-
-      // Scenario A: Merge Restore (Simulating Sync download phase)
-      await database.importBackupData(backupPayload, "merge");
-      const afterMerge = await database.getAllCards();
-
-      // Reset for Scenario B: Replace Restore (Simulating Force Recovery)
-      await resetData();
-      await database.styleCards.bulkAdd([
-        {
-          id: "card-1",
-          name: "Local Card 1 (Old)",
-          createdAt: 1000,
-          updatedAt: 1000,
-          promptSegments: [],
-          parameters: {},
-          masking: {},
-          tier: "Common",
-          tags: [],
-          dominantColor: "#000",
-          thumbnailData: ""
-        },
-        {
-          id: "card-2",
-          name: "Local Only Card",
-          createdAt: 1000,
-          updatedAt: 1000,
-          promptSegments: [],
-          parameters: {},
-          masking: {},
-          tier: "Common",
-          tags: [],
-          dominantColor: "#000",
-          thumbnailData: ""
-        }
-      ]);
-
-      await database.importBackupData(backupPayload, "replace");
-      const afterReplace = await database.getAllCards();
-
-      // Scenario C: Logical Delete Sync
-      await resetData();
-      await database.styleCards.bulkAdd([
-        {
-          id: "card-active",
-          name: "Active Card",
-          createdAt: 1000,
-          updatedAt: 1000,
-          promptSegments: [],
-          parameters: {},
-          masking: {},
-          tier: "Common",
-          tags: [],
-          dominantColor: "#000",
-          thumbnailData: "heavy-thumb-data"
-        }
-      ]);
-
-      const backupWithDelete = {
-        styleCards: [
-          {
-            id: "card-active",
-            name: "Active Card",
-            createdAt: 1000,
-            updatedAt: 2000, // Newer delete state
-            isDeleted: true,
-            promptSegments: [],
-            parameters: {},
-            masking: {},
-            tier: "Common",
-            tags: [],
-            dominantColor: "#000",
-            thumbnailData: "",
-            images: [],
-            selectedThumbnails: []
-          }
-        ],
-        categories: [],
-        userSettings: [],
-        historyItems: []
-      };
-
-      await database.importBackupData(backupWithDelete, "merge");
-      const activeCardsAfterDeleteSync = await database.getAllCards();
-      const physicalCardsAfterDeleteSync = await database.styleCards.toArray();
-
-      return {
-        afterMerge: afterMerge.map((c: any) => ({ id: c.id, name: c.name })),
-        afterReplace: afterReplace.map((c: any) => ({ id: c.id, name: c.name })),
-        logicalDeleteSync: {
-          activeCount: activeCardsAfterDeleteSync.length,
-          physicalCount: physicalCardsAfterDeleteSync.length,
-          card: physicalCardsAfterDeleteSync[0] ? {
-            id: physicalCardsAfterDeleteSync[0].id,
-            isDeleted: physicalCardsAfterDeleteSync[0].isDeleted,
-            thumbnailData: physicalCardsAfterDeleteSync[0].thumbnailData
-          } : null
-        }
-      };
-    });
-
-    console.log("Merge results:", results.afterMerge);
-    console.log("Replace results:", results.afterReplace);
-    console.log("Delete sync results:", results.logicalDeleteSync);
-
-    // Verify Merge Results
-    expect(results.afterMerge).toContainEqual({ id: "card-1", name: "Backup Card 1 (New)" });
-    expect(results.afterMerge).toContainEqual({ id: "card-2", name: "Local Only Card" });
-    expect(results.afterMerge).toContainEqual({ id: "card-3", name: "Backup Only Card" });
-    expect(results.afterMerge.length).toBe(3);
-
-    // Verify Replace Results
-    expect(results.afterReplace).toContainEqual({ id: "card-1", name: "Backup Card 1 (New)" });
-    expect(results.afterReplace).toContainEqual({ id: "card-3", name: "Backup Only Card" });
-    expect(results.afterReplace.length).toBe(2);
-
-    // Verify Logical Delete Sync Results
-    expect(results.logicalDeleteSync.activeCount).toBe(0); // Filtered out from getAllCards
-    expect(results.logicalDeleteSync.physicalCount).toBe(1); // Record still exists in IndexedDB
-    expect(results.logicalDeleteSync.card).toEqual({
-      id: "card-active",
-      isDeleted: true,
-      thumbnailData: "" // Heavy image info is cleared
-    });
-
-    console.log("Sync and logical delete E2E logic verification passed successfully!");
-  });
-
-  test("should toggle Easy Mode and restrict tab visibility to Library only", async ({ page }) => {
-    const screenshotsDir = path.join(__dirname, "../../tests/screenshots");
-    console.log("Navigating to sandbox page for Easy Mode E2E test...");
-    await page.goto("/tests/sandbox/index.html");
-
-    const spFrame = page.frameLocator("#sidepanel-frame");
-
-    // 1. Skip welcome dialog if exists
-    const skipButton = spFrame.locator("text=スキップ");
-    if (await skipButton.isVisible({ timeout: 5000 }).catch(() => false)) {
-      await skipButton.click();
-    }
-
-    // 2. Verify History tab is visible initially
-    const historyTabBtn = spFrame.locator("nav button:has-text('History')");
-    await expect(historyTabBtn).toBeVisible({ timeout: 10000 });
-
-    // 3. Click Settings icon in header to navigate to Settings
-    const settingsNavBtn = spFrame.locator("#settings-nav-btn");
-    await expect(settingsNavBtn).toBeVisible({ timeout: 10000 });
-    await settingsNavBtn.click();
-
-    // 4. Verify Easy Mode toggle button exists in SettingsTab
-    const easyModeToggle = spFrame.locator("#easy-mode-toggle-btn");
-    await expect(easyModeToggle).toBeVisible({ timeout: 10000 });
-
-    // 5. Toggle Easy Mode to ON
-    console.log("Enabling Easy Mode...");
-    await easyModeToggle.click();
-    await page.waitForTimeout(500);
-
-    // 6. Verify tabs are hidden and Library title is displayed
-    const libraryTitle = spFrame.locator("span:has-text('Library')");
-    await expect(libraryTitle).toBeVisible({ timeout: 10000 });
-    await expect(historyTabBtn).not.toBeVisible();
-
-    // 7. Save screenshot of active Easy Mode
-    await page.screenshot({
-      path: path.join(screenshotsDir, "easy-mode-active.png"),
-    });
-    console.log("Easy Mode active screenshot saved.");
-
-    // 8. Re-open settings via header settings button
-    await settingsNavBtn.click();
-    await page.waitForTimeout(500);
-
-    // 9. Toggle Easy Mode to OFF
-    console.log("Disabling Easy Mode...");
-    await easyModeToggle.click();
-    await page.waitForTimeout(500);
-
-    // 10. Verify typical tabs are restored (e.g. History tab visible)
-    await expect(historyTabBtn).toBeVisible({ timeout: 10000 });
-
-    // 11. Save screenshot of restored standard mode
-    await page.screenshot({
-      path: path.join(screenshotsDir, "easy-mode-deactivated.png"),
-    });
-    console.log("Easy Mode E2E test passed successfully!");
-  });
-});
